requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.7
aiolifxc==1.0.0
aiosignal==1.3.1
APScheduler==3.10.4
attrs==24.2.0
bitarray==2.9.3
bitstring==4.2.3
cbor==1.0.0
click==8.1.7
click-log==0.4.0
contourpy==1.3.1
coppeliasim_zmqremoteapi_client==2.0.4
cycler==0.12.1
docopt==0.6.2
fonttools==4.55.3
frozenlist==1.5.0
hbmqtt==0.9.6
idna==3.10
kiwisolver==1.4.8
matplotlib==3.10.0
multidict==6.1.0
numpy==2.2.1
opencv-python==4.10.0.84
packaging==24.2
passlib==1.7.4
pillow==11.1.0
propcache==0.2.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
scipy==1.15.0
six==1.16.0
transitions==0.9.2
tzlocal==5.2
websockets==14.1
yarl==1.18.0
zmq==0.0.0
